Fix q&a bugs
